remove support for boolean fn validators
the problem runs too deep at the moment, with a fn pointer type for closure such as |a| a> 0 its not possible to make it work with both value and ref type (e.g. i32 and String) because its not possible to abstract over ref.

This could be resolved by just supplying e.g. `val > 0` expression with the contract being that `val` will be bound in the outer generated code to the field value. This way the generated code would essentially be same as in the Closure case with only the comparison expression for the if being supplied by the user, but this is very hard to check and so its very easy to get some bad errors.